+++ CHANGES.md
@@ -4,18 +4,22 @@
 ### Added
 - `guard::GuardContext` for use with the `Guard` trait. [#2552]
 - `ServiceRequest::guard_ctx` for obtaining a guard context. [#2552]
+- `impl Hash` for `http::header::Encoding`. [#2501]
+- `AcceptEncoding::negotiate`. [#2501]
 
 ### Changed
 - `Guard` trait now receives a `&GuardContext`. [#2552]
 - `guard::fn_guard` functions now receives a `&GuardContext`. [#2552]
 - Some guards now return `impl Guard` and their concrete types are made private: `guard::{Header}` and all the method guards. [#2552]
 - The `Not` guard is now generic over the type of guard it wraps. [#2552]
+- `AcceptEncoding::preference` now returns `Option<Preference<Encoding>>`. [#2501]
 
 ### Fixed
 - Rename `ConnectionInfo::{remote_addr => peer_addr}`, deprecating the old name. [#2554]
 - `ConnectionInfo::peer_addr` will not return the port number. [#2554]
 - `ConnectionInfo::realip_remote_addr` will not return the port number if sourcing the IP from the peer's socket address. [#2554]
 
+[#2501]: https://github.com/actix/actix-web/pull/2501
 [#2552]: https://github.com/actix/actix-web/pull/2552
 [#2554]: https://github.com/actix/actix-web/pull/2554